requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-rest-swagger==2.2.0
djangorestframework==3.11.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.5
pytz==2020.1
ratelimit==2.2.1
redis==3.5.3
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
structlog==20.1.0
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
